Move tests to test_spans
